BOT/requirements.txt
discord.py==1.7.3
requests==2.24.0
pandas==1.4.1